Created ClaimStatusDisplay and updated UI to reflect
