+++ src/Core/Core.csproj
@@ -42,7 +42,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="AppRoles.cs" />
-    <Compile Include="ClaimStatus.cs" />
+    <Compile Include="ClaimStatusDisplay.cs" />
     <Compile Include="IProfile.cs" />
     <Compile Include="IValidator.cs" />
     <Compile Include="Constants.cs" />
@@ -56,7 +56,7 @@
     <Compile Include="IRepository.cs" />
     <Compile Include="IUnitOfWork.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Status.cs" />
+    <Compile Include="ClaimStatus.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
