--- src/Core/Status.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace EDeviceClaims.Core
-{
-    public enum Status
-    {
-        Open,
-        Denied,
-        Approved
-    }
-}
